backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]>=2.0.40
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# AI / LLM
openai>=2.0.0

# Document parsing
pypdf>=4.3.0
python-docx>=1.1.0
python-multipart==0.0.12

# Task queue
celery==5.4.0
redis==5.2.1

# Observability
langfuse>=4.0.0

# Web search
tavily-python==0.5.0

# Validation
pydantic>=2.11.0
pydantic-settings>=2.8.0

# Utilities
python-dotenv==1.0.1
httpx>=0.27.0
tenacity==9.0.0
numpy>=1.26.4,<2.0
